google-docs-cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gdocs-cli"
version = "0.1.0"
description = "A CLI tool for interacting with Google Docs"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "click>=8.1.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.0",
    "rich>=13.0.0",
]

[project.scripts]
gdocs = "gdocs_cli.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
